+++ app/mailers/user_mailer.rb
@@ -2,36 +2,32 @@ class UserMailer < ActionMailer::Base
   default from: Cartodb.get_config(:mailer, 'from')
   layout 'mail'
 
-  GREETINGS = %w(congrats congratulations cool awesome hooray nice wow rad bravo yay boom).freeze
-
   def new_organization_user(user)
     @user = user
     @organization = @user.organization
     @owner = @organization.owner
     @subject = "You have been invited to CARTO organization '#{@organization.name}'"
-    base_url = CartoDB.base_url(@organization.name, @user.username)
 
     if @user.enable_account_token.nil?
-      @link = base_url
+      @link = "#{CartoDB.base_url(@organization.name, @user.username)}"
     else
-      path = CartoDB.path(self, 'enable_account_token_show', id: @user.enable_account_token)
-      @enable_account_link = "#{base_url}#{path}"
+      @enable_account_link = "#{CartoDB.base_url(@organization.name, @user.username)}#{CartoDB.path(self, 'enable_account_token_show', {id: @user.enable_account_token})}"
     end
 
     @user_needs_password = user_needs_password(user)
 
-    mail to: @user.email, subject: @subject
+    mail :to => @user.email,
+         :subject => @subject
   end
 
   def share_table(table, user)
     @table_visualization = table
     @user = user
     organization = @table_visualization.user.organization
-    visualization_username = @table_visualization.user.username
-    path = CartoDB.path(self, 'public_tables_show_bis', id: "#{visualization_username}.#{@table_visualization.name}")
-    @link = "#{CartoDB.base_url(organization.name, @user.username)}#{path}"
-    @subject = "#{visualization_username} has shared a CARTO dataset with you"
-    mail to: @user.email, subject: @subject
+    @link = "#{CartoDB.base_url(organization.name, @user.username)}#{CartoDB.path(self, 'public_tables_show_bis', {id: "#{@table_visualization.user.username}.#{@table_visualization.name}"})}"
+    @subject = "#{@table_visualization.user.username} has shared a CARTO dataset with you"
+    mail :to => @user.email,
+         :subject => @subject
   end
 
   def share_visualization(visualization, user)
@@ -42,7 +38,8 @@ class UserMailer < ActionMailer::Base
     visualization_presenter = Carto::Api::VisualizationPresenter.new(visualization, user, self)
     @link = visualization_presenter.privacy_aware_map_url
     @subject = "#{@visualization.user.username} has shared a CARTO map with you"
-    mail to: @user.email, subject: @subject
+    mail(to: @user.email,
+         subject: @subject)
   end
 
   def unshare_table(table_visualization_name, table_visualization_owner_name, user)
@@ -50,7 +47,8 @@ class UserMailer < ActionMailer::Base
     @table_visualization_owner_name = table_visualization_owner_name
     @user = user
     @subject = "#{@table_visualization_owner_name} has stopped sharing a CARTO dataset with you"
-    mail to: @user.email, subject: @subject
+    mail :to => @user.email,
+         :subject => @subject
   end
 
   def unshare_visualization(visualization_name, visualization_owner_name, user)
@@ -58,7 +56,8 @@ class UserMailer < ActionMailer::Base
     @visualization_owner_name = visualization_owner_name
     @user = user
     @subject = "#{@visualization_owner_name} has stopped sharing a CARTO map with you"
-    mail to: @user.email, subject: @subject
+    mail :to => @user.email,
+         :subject => @subject
   end
 
   def map_liked(visualization, viewer_user, visualization_preview_image)
@@ -67,12 +66,12 @@ class UserMailer < ActionMailer::Base
     @viewer_name = viewer_user.name_or_username
     @preview_image = visualization_preview_image
     @subject = "Your map got some love!"
-    @greetings = GREETINGS
+    @greetings = ["congrats", "congratulations", "cool", "awesome", "hooray", "nice", "wow", "rad", "bravo", "yay", "boom"]
     mail_tracker = get_mail_tracker('like_map')
-    path = CartoDB.path(self, 'public_visualizations_show_map', id: visualization.id)
-    @link = "#{@user.public_url}#{path}#{mail_tracker}"
+    @link = "#{@user.public_url}#{CartoDB.path(self, 'public_visualizations_show_map', id: visualization.id)}#{mail_tracker}"
     @viewer_maps_link = "#{viewer_user.public_url}#{CartoDB.path(self, 'public_maps_home')}"
-    mail to: @user.email, subject: @subject
+    mail :to => @user.email,
+         :subject => @subject
   end
 
   def table_liked(canonical_visualization, viewer_user, visualization_preview_image)
@@ -81,12 +80,12 @@ class UserMailer < ActionMailer::Base
     @viewer_name = viewer_user.name_or_username
     @preview_image = visualization_preview_image
     @subject = "Your dataset got some love!"
-    @greetings = GREETINGS
+    @greetings = ["congrats", "congratulations", "cool", "awesome", "hooray", "nice", "wow", "rad", "bravo", "yay", "boom"]
     mail_tracker = get_mail_tracker('like_map')
-    path = CartoDB.path(self, 'public_visualizations_show', id: canonical_visualization.id)
-    @link = "#{@user.public_url}#{path}#{mail_tracker}"
+    @link = "#{@user.public_url}#{CartoDB.path(self, 'public_visualizations_show', id: canonical_visualization.id)}#{mail_tracker}"
     @viewer_datasets_link = "#{viewer_user.public_url}#{CartoDB.path(self, 'public_datasets_home')}"
-    mail to: @user.email, subject: @subject
+    mail :to => @user.email,
+         :subject => @subject
   end
 
   def trending_map(visualization, mapviews, visualization_preview_image)
@@ -95,21 +94,11 @@ class UserMailer < ActionMailer::Base
     @map_name = visualization.name
     @preview_image = visualization_preview_image
     @subject = "Recent activity on one of your maps!"
-    @greetings = GREETINGS
+    @greetings = ["congrats", "congratulations", "cool", "awesome", "hooray", "nice", "wow", "rad", "bravo", "yay", "boom"]
     mail_tracker = get_mail_tracker('trending_map')
-    path = CartoDB.path(self, 'public_visualizations_show_map', id: visualization.id)
-    @link = "#{@user.public_url}#{path}#{mail_tracker}"
-    mail to: @user.email, subject: @subject
-  end
-
-  def password_reset(user)
-    @user = user
-    @organization = @user.organization
-    @subject = "Reset CARTO password"
-    base_url = CartoDB.base_url(@organization.try(:name))
-    path = CartoDB.path(self, 'edit_password_reset', id: @user.password_reset_token)
-    @password_reset_link = "#{base_url}#{path}"
-    mail to: @user.email, subject: @subject
+    @link = "#{@user.public_url}#{CartoDB.path(self, 'public_visualizations_show_map', id: visualization.id)}#{mail_tracker}"
+    mail :to => @user.email,
+         :subject => @subject
   end
 
   private