+++ app/models/carto/feature_flag.rb
@@ -1,15 +1,17 @@
 # encoding: UTF-8
 
 module Carto
-  class FeatureFlag < ActiveRecord::Base
 
+  class FeatureFlag < ActiveRecord::Base
+   
     validates :name, presence: true
 
-    has_many :feature_flags_user, dependent: :destroy
+    has_many :feature_flags_user, :dependent => :destroy
 
     def self.find_by_user(user)
-      FeatureFlag.where(restricted: false) + user.feature_flags.select(&:restricted)
+      FeatureFlag.where(restricted: false) + user.feature_flags.select { |feature_flag| feature_flag.restricted }
     end
 
   end
+
 end