+++ lib/resque/reporter_jobs.rb
@@ -1,5 +1,4 @@
 # encoding: utf-8
-
 require_relative './base_job'
 
 module Resque
@@ -9,7 +8,7 @@ module Resque
         @queue = :users
 
         def self.perform(mail_to, trending_visualizations)
-          ReporterMailer.trending_maps_report(mail_to, trending_visualizations).deliver_now
+          ReporterMailer.trending_maps_report(mail_to, trending_visualizations).deliver
         end
       end
     end